+++ tests/integration/integrations/test_loki.py
@@ -19,12 +19,12 @@ logger = logging.getLogger(__name__)
 @pytest.mark.parametrize(
     "app_fixture, port",
     [
+        ("flask_app", 8000),
+        ("django_app", 8000),
         ("spring_boot_app", 8080),
         ("expressjs_app", 8080),
         ("go_app", 8080),
         ("fastapi_app", 8080),
-        ("flask_app", 8000),
-        ("django_app", 8000),
     ],
 )
 def test_loki_integration(