+++ tests/integration/integrations/test_oauth.py
@@ -20,12 +20,12 @@ logger = logging.getLogger(__name__)
 @pytest.mark.parametrize(
     "app_fixture, endpoint",
     [
+        ("flask_app", "login"),
+        ("django_app", "auth_login"),
         ("expressjs_app", "login"),
         ("go_app", "login/openid-connect"),
         ("fastapi_app", "login"),
         ("spring_boot_app", "oauth2/authorization/oidc"),
-        ("flask_app", "login"),
-        ("django_app", "auth_login"),
     ],
 )
 def test_oauth_integrations(