+++ tests/integration/integrations/test_openfga.py
@@ -17,11 +17,11 @@ logger = logging.getLogger(__name__)
 @pytest.mark.parametrize(
     "app_fixture, port",
     [
-        ("spring_boot_app", 8080),
         ("flask_app", 8000),
         ("django_app", 8000),
         ("fastapi_app", 8080),
         ("go_app", 8080),
+        ("spring_boot_app", 8080),
     ],
 )
 def test_openfga_integrations(