+++ docs/changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Place any unreleased changes here, that are subject to release in coming versions :).
 
 * docs: Added release notes for `paas-charm v1.9`.
+* fix: Fixed FastAPI example and changed to use edge PostgreSQL for integration tests.
 
 ## 1.9.0 - 2025-10-07